moved func.py in func package
