+++ run.py
@@ -1,5 +1,5 @@
 from src.trudvsem import TrudVsem
-from src.func import run_user_interface
+from src.func.func import run_user_interface
 
 
 if __name__ == "__main__":
